+++ R/PKNCA.options.R
@@ -223,7 +223,7 @@
         start=0,
         end=c(24, Inf),
         auclast=c(TRUE, FALSE),
-        aucinf=c(FALSE, TRUE),
+        aucinf.obs=c(FALSE, TRUE),
         half.life=c(FALSE, TRUE),
         tmax=c(FALSE, TRUE),
         cmax=c(FALSE, TRUE))